stlport conformity
git-svn-id: https://surf.svn.sourceforge.net/svnroot/surf/trunk@207 5f8b00c5-0c0e-0410-9123-c6bd0749f4b9

+++ surf/image-formats/PNG.cc
@@ -106,7 +106,7 @@ namespace ImageFormats {
 
 		png_bytep ptr = const_cast<png_bytep>(buffer.getMap());
 		png_bytepp row_pointers = new png_bytep[height];
-		for(size_t i = 0; i != height; i++) {
+		for(int i = 0; i != height; i++) {
 			row_pointers[i] = ptr;
 			ptr += width;
 		}
